Move client/lib/stories to client/stories
