+++ client/.storybook/webpack.config.js
@@ -11,6 +11,7 @@ module.exports = ({ config }) => {
   })
 
   config.resolve.alias['~'] = path.join(__dirname, '..', 'lib')
+  config.resolve.alias['~stories'] = path.join(__dirname, '..', 'stories')
   config.resolve.extensions.push('.ts', '.tsx')
 
   return config
